+++ NEWS.md
@@ -1,4 +1,4 @@
-# waldo (development version)
+# waldo 0.3.1
 
 * `compare()`ing data frames now works independently of `option(max.print)` 
   (#105).